requirements.txt
# Core scientific stack
numpy>=1.26
scipy>=1.11
pandas>=2.2
scikit-learn>=1.4
tqdm>=4.66
anndata>=0.10
scanpy>=1.10
scikit-misc>=0.3
python-igraph>=0.11
leidenalg>=0.10

# Model / training
torch>=2.2

# LLM / config
openai>=2.24
python-dotenv>=1.0
pydantic>=2.7
pydantic-settings>=2.11

# Web portal
fastapi>=0.128
uvicorn>=0.37
